move text box tweaked
