+++ scripts/screens/EventEditScreen.py
@@ -22,7 +22,6 @@ from scripts.cat.pronouns import get_default_pronouns
 from scripts.game_structure.screen_settings import MANAGER
 from scripts.game_structure.ui_elements import (
     UIModifiedScrollingContainer,
-    UITextBoxTweaked,
     UICheckbox,
     UIModifiedImage,
     UIScrollingButtonList,
@@ -30,6 +29,7 @@ from scripts.game_structure.ui_elements import (
     UICollapsibleContainer,
     UIScrollingDropDown,
 )
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.windows.editor_save_check import EditorSaveCheck
 from scripts.ui.windows.editor_missing_info import EditorMissingInfoWindow
