+++ scripts/screens/MedDenScreen.py
@@ -8,9 +8,9 @@ from scripts.cat.cats import Cat
 from scripts.clan_resources.herb.herb_supply import MESSAGES
 from scripts.game_structure import game
 from scripts.game_structure.ui_elements import (
-    UITextBoxTweaked,
     UIModifiedImage,
 )
+from ..ui.elements.text_box_tweaked import UITextBoxTweaked
 from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
