+++ scripts/screens/RoleScreen.py
@@ -8,9 +8,7 @@ import pygame_gui
 
 from scripts.cat.cats import Cat
 from scripts.game_structure import image_cache, game
-from scripts.game_structure.ui_elements import (
-    UITextBoxTweaked,
-)
+from ..ui.elements.text_box_tweaked import UITextBoxTweaked
 from ..ui.elements.surface_image_button import UISurfaceImageButton
 from ..ui.theme import get_text_box_theme
 from ..events_module.text_adjust import adjust_list_text, shorten_text_to_fit
