+++ scripts/ui/windows/delete_check.py
@@ -4,9 +4,7 @@ import shutil
 import pygame
 import pygame_gui
 
-from scripts.game_structure.ui_elements import (
-    UITextBoxTweaked,
-)
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.housekeeping.datadir import get_save_dir
 from scripts.screens.enums import GameScreen
