+++ scripts/ui/windows/editor_missing_info.py
@@ -1,5 +1,5 @@
 import pygame
-from scripts.game_structure.ui_elements import UITextBoxTweaked
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.ui.windows.window_base_class import GameWindow
 from scripts.ui.scale import ui_scale
