+++ scripts/ui/windows/editor_save_check.py
@@ -2,9 +2,7 @@ import pygame
 import pygame_gui
 import ujson
 
-from scripts.game_structure.ui_elements import (
-    UITextBoxTweaked,
-)
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.windows.window_base_class import GameWindow
