+++ scripts/ui/windows/update.py
@@ -3,9 +3,7 @@ import threading
 import pygame
 import pygame_gui
 
-from scripts.game_structure.ui_elements import (
-    UITextBoxTweaked,
-)
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 from scripts.housekeeping.progress_bar_updater import UIUpdateProgressBar
 from scripts.housekeeping.update import self_update, UpdateChannel
 from scripts.housekeeping.version import get_version_info
